chore: move IntegrationTest and improve comments
TIS21-5226: Disconnect Doctors
